// packages/front/src/pages/my-music/types.ts
import { baseTypes } from "@vk-audiopad/common";
import { TCoverPlaylist } from "shared/types";

export type TFetchMyMusicResult = {
    playlist: baseTypes.TTitlePlaylist | null;
    recentTracksPlaylist: baseTypes.TTitlePlaylist | null;
    radiostationsPlaylist: baseTypes.TTitlePlaylist | null;
    coverPlaylists: TCoverPlaylist[];
};